adding bnl condor megamixer
This has to work on more constraints than the
others.

for example, we cannot write log files to the output
location, too many open files on gpfs can cause problems.

Going straight to the condor scratch dir, and need
to write all files there first, other than the
fits files which are not held open.

Also the scripts need to know where one another are. So
using absolute paths, different from what Matt does. Might
be able to adapt this to relative paths by getting the
absolute paths to things and storing them in a variable