deal with different models from nbrs
We can fit a different model that used in MOF.

This includes different psf models.

I now require keywords in the config specifying the models used for MOF.
This is a stopgap, we should somehow retrive this info from the
specified mof run.  Need the install command to copy the config I think.

Also don't use the central as a nbr of itself, since the psf could be
different.